+++ src/modules/grid.module.ts
@@ -4,7 +4,7 @@ import {CommonModule} from "@anglr/common";
 
 import {GridComponent} from "../components/grid/grid.component";
 import {BasicPagingComponent, LoadMorePagingComponent, PreviousNextPagingComponent, ContentVirtualScrollPagingComponent, PageVirtualScrollPagingComponent, NoPagingComponent} from "../plugins/paging/components";
-import {NoGridInitializerComponent, QueryCookieGridInitializerComponent, QueryGridInitializerComponent} from "../plugins/gridInitializer/components";
+import {NoGridInitializerComponent, QueryPermanentStorageGridInitializerComponent, QueryGridInitializerComponent} from "../plugins/gridInitializer/components";
 import {AsyncDataLoaderComponent, SyncDataLoaderComponent} from "../plugins/dataLoader/components";
 import {BasicTableMetadataGathererComponent} from "../components/metadata/basicTable/basicTableMetadataGatherer.component";
 import {BasicTableColumnComponent} from "../components/metadata/basicTable/basicTableColumn.component";
@@ -35,7 +35,7 @@ import {BasicRowSelectorComponent, LimitedRowSelectorComponent} from "../plugins
         PreviousNextPagingComponent,
         NoPagingComponent,
         QueryGridInitializerComponent,
-        QueryCookieGridInitializerComponent,
+        QueryPermanentStorageGridInitializerComponent,
         NoGridInitializerComponent,
         SyncDataLoaderComponent,
         AsyncDataLoaderComponent,
@@ -61,7 +61,7 @@ import {BasicRowSelectorComponent, LimitedRowSelectorComponent} from "../plugins
         PreviousNextPagingComponent,
         NoPagingComponent,
         QueryGridInitializerComponent,
-        QueryCookieGridInitializerComponent,
+        QueryPermanentStorageGridInitializerComponent,
         NoGridInitializerComponent,
         SyncDataLoaderComponent,
         AsyncDataLoaderComponent,