+++ src/plugins/gridInitializer/components.ts
@@ -1,3 +1,3 @@
 export * from './no/noGridInitializer.component';
 export * from './query/queryGridInitializer.component';
-export * from './queryCookie/queryCookieGridInitializer.component';
+export * from './queryPermanentStorage/queryPermanentStorageGridInitializer.component';