+++ src/plugins/gridInitializer/index.ts
@@ -1,4 +1,4 @@
 export * from './gridInitializer.interface';
 export * from './no/noGridInitializer.interface';
 export * from './query/queryGridInitializer.interface';
-export * from './queryCookie/queryCookieGridInitializer.interface';
+export * from './queryPermanentStorage/queryPermanentStorageGridInitializer.interface';