+++ src/plugins/gridInitializer/queryPermanentStorage/queryPermanentStorageGridInitializer.component.ts
@@ -1,53 +1,53 @@
 import {ChangeDetectionStrategy, Inject, Optional, Component} from "@angular/core";
 import {ActivatedRoute, Router} from "@angular/router";
-import {CookieService} from "@anglr/common";
+import {PermanentStorage, PERMANENT_STORAGE} from "@anglr/common";
 import {extend, isPresent} from "@jscrpt/common";
 
 import {GridPluginGeneric} from "../../../misc";
 import {GRID_INITIALIZER_OPTIONS} from "../types";
-import {QueryCookieGridInitializerOptions, QueryCookieGridInitializer} from "./queryCookieGridInitializer.interface";
+import {QueryPermanentStorageGridInitializerOptions, QueryPermanentStorageGridInitializer} from "./queryPermanentStorageGridInitializer.interface";
 import {QueryGridInitializerComponent} from "../query/queryGridInitializer.component";
 
 /**
- * Component used for rendering query cookie grid initializer
+ * Component used for rendering query, permanent storage grid initializer
  */
 @Component(
 {
-    selector: "ng-query-cookie-grid-initializer",
+    selector: "ng-query-permanent-storage-grid-initializer",
     template: '',
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class QueryCookieGridInitializerComponent extends QueryGridInitializerComponent implements QueryCookieGridInitializer, GridPluginGeneric<QueryCookieGridInitializerOptions>
+export class QueryPermanentStorageGridInitializerComponent extends QueryGridInitializerComponent implements QueryPermanentStorageGridInitializer, GridPluginGeneric<QueryPermanentStorageGridInitializerOptions>
 {
     //######################### public properties - implementation of NoGridInitializer #########################
 
     /**
      * Options for grid plugin
      */
-    public get options(): QueryCookieGridInitializerOptions
+    public get options(): QueryPermanentStorageGridInitializerOptions
     {
         return this._options;
     }
-    public set options(options: QueryCookieGridInitializerOptions)
+    public set options(options: QueryPermanentStorageGridInitializerOptions)
     {
-        this._options = extend(true, this._options, options) as QueryCookieGridInitializerOptions;
+        this._options = extend(true, this._options, options) as QueryPermanentStorageGridInitializerOptions;
     }
 
     //######################### protected properties #########################
 
     /**
-     * Gets name of items per page cookie
+     * Gets name of items per page permanent storage
      */
-    protected get itemsPerPageCookieName(): string
+    protected get itemsPerPagePermanentStorageName(): string
     {
-        return this.options.cookieIppName;
+        return this.options.storageIppName;
     }
 
     //######################### constructor #########################
     constructor(_router: Router,
                 _route: ActivatedRoute,
-                @Optional() protected _cookies: CookieService,
-                @Inject(GRID_INITIALIZER_OPTIONS) @Optional() options?: QueryCookieGridInitializerOptions)
+                @Inject(PERMANENT_STORAGE) @Optional() protected _permanentStorages: PermanentStorage,
+                @Inject(GRID_INITIALIZER_OPTIONS) @Optional() options?: QueryPermanentStorageGridInitializerOptions)
     {
         super(_router, _route, options);
     }
@@ -66,9 +66,9 @@ export class QueryCookieGridInitializerComponent extends QueryGridInitializerCom
             return ipp;
         }
 
-        if(this._cookies && this.itemsPerPageCookieName)
+        if(this._permanentStorages && this.itemsPerPagePermanentStorageName)
         {
-            ipp = this._cookies.getCookie(this.itemsPerPageCookieName) as number;
+            ipp = this._permanentStorages.get<number>(this.itemsPerPagePermanentStorageName);
 
             if(isPresent(ipp))
             {
@@ -85,9 +85,9 @@ export class QueryCookieGridInitializerComponent extends QueryGridInitializerCom
      */
     public setItemsPerPage(itemsPerPage: number)
     {
-        if(this._cookies && this.itemsPerPageCookieName)
+        if(this._permanentStorages && this.itemsPerPagePermanentStorageName)
         {
-            this._cookies.setCookie(this.itemsPerPageCookieName, itemsPerPage, null, "/");
+            this._permanentStorages.set(this.itemsPerPagePermanentStorageName, itemsPerPage);
         }
 
         super.setItemsPerPage(itemsPerPage);